ORV2-1562 Fin Report Refactoring and query logic
